Guard against loops in keyboard hook
If the keyboard hook handler sends the same key that is being
reacted to, the handler will loop with itself. This change
guards against that problem by ignoring virtual key presses.

You see this problem when you quickly release the VimKey. It takes
a second or so before the key appears because it is stuck in a loop.
Now that the loop is prevented, the key appears immediately.