Continue holding recent modifiers on vim key release
Normally, a keypress is sent as soon as a key is pressed down.
However, the modifier key was changed to only send the keypress
once the key has been released. This causes issues when the
user releases a modifier key before releasing the vim key,
because then the normal key is signalled instead of the intended
modified key.
The solution in this commit is to pretend any recently released
modifier keys are still held down when the vim key is released.